@@ -8,23 +8,22 @@
     ],
     "license": "MIT",
     "require": {
-        "php": "^7.2.5",
+        "php": "^7.3",
         "fideloper/proxy": "^4.2",
         "fruitcake/laravel-cors": "^2.0",
-        "guzzlehttp/guzzle": "^7.0",
-        "laravel/framework": "^7.28",
+        "guzzlehttp/guzzle": "^7.0.1",
+        "laravel/framework": "^8.13",
         "laravel/tinker": "^2.0",
-        "laravel/ui": "^2.0",
-        "spatie/laravel-permission": "^3.16"
+        "laravel/ui": "^3.0",
+        "spatie/laravel-permission": "^3.17"
     },
     "require-dev": {
-        "facade/ignition": "^2.0",
+        "facade/ignition": "^2.3.6",
         "fzaninotto/faker": "^1.9.1",
         "jasonmccreary/laravel-test-assertions": "^1.0",
         "mockery/mockery": "^1.3.1",
-        "mpociot/laravel-test-factory-helper": "^2.1",
-        "nunomaduro/collision": "^4.1",
-        "phpunit/phpunit": "^8.5"
+        "nunomaduro/collision": "^5.0",
+        "phpunit/phpunit": "^9.3"
     },
     "config": {
         "optimize-autoloader": true,
@@ -38,12 +37,10 @@
     },
     "autoload": {
         "psr-4": {
-            "App\\": "app/"
-        },
-        "classmap": [
-            "database/seeds",
-            "database/factories"
-        ]
+            "App\\": "app/",
+            "Database\\Factories\\": "database/factories/",
+            "Database\\Seeders\\": "database/seeders/"
+        }
     },
     "autoload-dev": {
         "psr-4": {
